refactor: Update package structure and logging configuration - Move security classes to appropriate packages - Enhance logging for exception handling - Clean up CORS configuration
