+++ pom.xml
@@ -51,6 +51,8 @@
             <artifactId>spring-boot-starter-data-jpa</artifactId>
         </dependency>
 
+        <!-- Actuator Eklenebilir -->
+
 
         <!-- PostgreSQL driver -->
         <dependency>
