+++ src/main/java/com/example/jwtsecurity/controller/BookingController.java
@@ -17,9 +17,10 @@ import java.util.Map;
 // rezervo için endpointler
 @RestController
 @RequestMapping("/api/bookings")
+// TODO: buralari security config altinda belirttiginde gerek olmamali orayi fixleyebiliriz bunu kaldirip
 @CrossOrigin(origins = {
     "http://localhost:3000",
-    "http://localhost:5173", 
+    "http://localhost:5173",
     "http://localhost:5174",
     "http://127.0.0.1:5173",
     "http://127.0.0.1:5174",
