+++ src/main/java/com/example/jwtsecurity/exception/GlobalExceptionHandler.java
@@ -1,5 +1,7 @@
 package com.example.jwtsecurity.exception;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
@@ -12,6 +14,10 @@ import java.util.Map;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
+
+    // Grafana, Graylog gibi sistemlere bu loglama aktarilabiliyor bir kac config ile
+    // Log4j logger implementation
+    private static final Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<Map<String, Object>> handleValidationExceptions(MethodArgumentNotValidException ex) {
@@ -35,12 +41,7 @@ public class GlobalExceptionHandler {
     public ResponseEntity<Map<String, Object>> handleGenericException(Exception ex) {
         Map<String, Object> response = new HashMap<>();
         response.put("error", true);
-        response.put("message", ex.getMessage());
-        response.put("exception", ex.getClass().getSimpleName());
-        
-        // debug için stack trace araştır ama
-        response.put("stackTrace", ex.getStackTrace());
-        
+        logger.error("Exception: ", ex);
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
     }
 }
