+++ src/main/java/com/example/jwtsecurity/service/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package com.example.jwtsecurity.service;
 
 import com.example.jwtsecurity.model.User;
 import com.example.jwtsecurity.repository.UserRepository;
-import com.example.jwtsecurity.security.CustomUserDetails;
+import com.example.jwtsecurity.model.CustomUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
