+++ src/main/resources/application.properties
@@ -29,7 +29,7 @@ jwt.secret=${JWT_SECRET:myVerySecretKeyThatShouldBeAtLeast256BitsLongForSecurity
 jwt.expiration=86400000
 
 # Spring Security Debug Logging
-logging.level.org.springframework.security=DEBUG
+logging.level.com.example.jwtsecurity=DEBUG
 logging.level.org.springframework.web.cors=DEBUG
 logging.level.org.springframework.security.web.FilterChainProxy=DEBUG
